% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistDistanceTree.R
\name{calculate_distance_matrix}
\alias{calculate_distance_matrix}
\title{Calculate Distance Matrix from Sequences}
\usage{
calculate_distance_matrix(sequences)
}
\arguments{
\item{sequences}{A list of character strings representing concert setlist sequences.}
}
\value{
A distance matrix where each entry (i, j) represents the distance between sequences i and j.
}
\description{
This function calculates a distance matrix based on the pairwise differences between sequences.
}
\examples{
seq_list <- list(showID1408 = "-5fD-UeE4OY-H]CJWhQla-mBk", showID1320 = "-5fD-^_E4OY-H]CJWhQeU-mBk", showID1336 = "5fD-UeE4OY-H]CJWhQei-mBk")
dist_matrix <- calculate_distance_matrix(seq_list)
}
