% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{codify_tour_song_titles}
\alias{codify_tour_song_titles}
\title{Codify Tour Song Titles}
\usage{
codify_tour_song_titles(concert_data)
}
\arguments{
\item{concert_data}{A data frame in concertData format.}
}
\value{
A data frame with columns:
  \itemize{
    \item song_title: The title of the song.
    \item four_letter_code: A unique 4-letter code for the song.
    \item hex_char: The unique hex character representation of the 4-letter code.
    \item hex_value: The unique hex value of the character.
  }
  Returns NULL if no songs are found in the dataset.
}
\description{
This function takes a concertData data frame, and for all
unique songs in the data set creates a mapping
with 4 columns: song title, 4-letter code, hex character, and hex value.
}
\examples{
\dontrun{
codified_songs <- codify_tour_song_titles(concert_data)
print(codified_songs)
}

}
