% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concertData_remove_showsNoSetlist}
\alias{concertData_remove_showsNoSetlist}
\title{Remove Shows with No Setlist from concertData}
\usage{
concertData_remove_showsNoSetlist(data)
}
\arguments{
\item{data}{A data frame in concertData format.}
}
\value{
A filtered concertData data frame with shows having no setlist removed.
}
\description{
Filters out shows with no setlist (no songs) from a concertData data frame.
}
\examples{
\dontrun{
filtered_data <- concertData_remove_showsNoSetlist(concert_data)
}

}
