% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concertData_remove_snippets}
\alias{concertData_remove_snippets}
\title{Remove Snippets from concertData}
\usage{
concertData_remove_snippets(data)
}
\arguments{
\item{data}{A data frame in concertData format.}
}
\value{
A filtered concertData data frame with snippet performances removed.
}
\description{
Filters out snippet performances from a concertData data frame.
}
\examples{
\dontrun{
filtered_data <- concertData_remove_snippets(concert_data)
}

}
