% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concertData_remove_tours}
\alias{concertData_remove_tours}
\title{Remove Specified Tours from concertData}
\usage{
concertData_remove_tours(data, excluded_tours)
}
\arguments{
\item{data}{A data frame in concertData format.}

\item{excluded_tours}{A character vector of tour names to be excluded.}
}
\value{
A filtered concertData data frame with specified tours removed.
}
\description{
Filters out specified tours from a concertData data frame.
}
\examples{
\dontrun{
filtered_data <- concertData_remove_tours(concert_data, c("Tour1", "Tour2"))
}

}
