% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{create_distinct_palette}
\alias{create_distinct_palette}
\title{Create a Palette of Distinct Colors}
\usage{
create_distinct_palette(n)
}
\arguments{
\item{n}{An integer specifying the number of distinct colors to generate.}
}
\value{
A character vector of hex color codes.
}
\description{
This function generates a palette of distinct colors that maximizes
differences between adjacent colors. It uses the HSV color space to create a
visually appealing and distinguishable set of colors.
}
\details{
Create a Palette of Distinct Colors


The function creates colors by:
\itemize{
  \item Using the full range of hues (0-1)
  \item Keeping saturation relatively high (0.7-1)
  \item Alternating value (brightness) between higher and lower values (0.6-0.9)
}
Hues are shuffled to avoid similar adjacent colors.
}
\examples{
\dontrun{
# Generate a palette of 10 distinct colors
colors <- create_distinct_palette(10)
# Use the colors in a plot
plot(1:10, 1:10, col = colors, pch = 19, cex = 3)
}

}
