% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistDistanceTree.R
\name{create_ggtree_plot}
\alias{create_ggtree_plot}
\title{Create a ggtree Plot}
\usage{
create_ggtree_plot(setlist_tree)
}
\arguments{
\item{setlist_tree}{A distance tree object to be plotted.}
}
\value{
A ggtree plot object.
}
\description{
This function generates a ggtree plot from a given distance tree.
}
\examples{
library(ggtree)
# Example tree structure (replace with actual tree object)
setlist_tree <- rtree(10)  # Generate a random tree with 10 tips
plot <- create_ggtree_plot(setlist_tree)
print(plot)
}
