% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{create_setlist_alignment}
\alias{create_setlist_alignment}
\title{Create Multiple Setlist Alignment}
\usage{
create_setlist_alignment(data, max_shows = 20)
}
\arguments{
\item{data}{A data frame containing at least the following columns:
\itemize{
  \item date: Date of the show
  \item song_title: Title of the song
}}

\item{max_shows}{Integer. Maximum number of representative shows to include in the alignment (default: 20)}
}
\value{
A data frame with aligned setlists, containing the following columns:
  \itemize{
    \item date: Date of the show
    \item position: Position of the song in the aligned setlist
    \item song_title: Title of the song (or NA for gaps in the alignment)
  }
}
\description{
This function creates a multiple setlist alignment using a custom
pairwise alignment approach followed the Needleman-Wunsch algorithm approach.
}
\examples{
\dontrun{
data(concert_data)
aligned_setlists <- create_setlist_alignment(concert_data, max_shows = 10)
head(aligned_setlists)
}

}
