% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{create_setlist_fasta_mafft}
\alias{create_setlist_fasta_mafft}
\title{Create FASTA Format for MAFFT Alignment}
\usage{
create_setlist_fasta_mafft(representative_setlists)
}
\arguments{
\item{representative_setlists}{A data frame containing two columns:
\itemize{
  \item showID: A character vector of show identifiers
  \item sequence: A character vector of encoded setlist sequences
}}
}
\value{
A character vector containing the FASTA format output, where each entry
consists of a header line (>showID) followed by the corresponding sequence.
}
\description{
This function creates a FASTA format output from representative setlists
for use in MAFFT multiple sequence alignment.
}
\examples{
\dontrun{
representative_setlists <- data.frame(
  showID = c("1408", "1320"),
  sequence = c("35 66 44 55 65 45 34 4F 59 48 5D 43 4A 57 68 51 6C 61 6D 42 6B 27",
               "35 66 44 5F 62 45 34 4F 59 48 5D 43 4A 57 68 51 60 69 6D 42 6B 27")
)
fasta_output <- create_setlist_fasta_mafft(representative_setlists)
print(fasta_output)
}

}
