% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourStatistics.R
\name{find_most_played_closing_songs}
\alias{find_most_played_closing_songs}
\title{Find Most Played Closing Songs}
\usage{
find_most_played_closing_songs(concert_data, n = 10)
}
\arguments{
\item{concert_data}{A data frame containing concert information with columns:
showID, song_title, song_position, and snippet}

\item{n}{Integer specifying how many top closing songs to return (default is 10)}
}
\value{
A data frame sorted by frequency (descending), containing columns:
        song_title and times_played (number of times played as closing song)
}
\description{
Identifies the most frequently played closing songs in concerts.
}
\examples{
\dontrun{
top_closers <- find_most_played_closing_songs(concert_data)
top_closers <- find_most_played_closing_songs(concert_data, n = 15)  # Show top 15
}

}
