% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourStatistics.R
\name{find_most_played_opening_songs}
\alias{find_most_played_opening_songs}
\title{Find Most Played Opening Songs}
\usage{
find_most_played_opening_songs(concert_data, n = NULL)
}
\arguments{
\item{concert_data}{A data frame containing concert information with columns:
date, city, song_title, song_position, and snippet}

\item{n}{Integer specifying how many top opening songs to return (default is all)}
}
\value{
A data frame sorted by frequency (descending), containing columns:
        song_title and times_played (number of times played as opening song)
}
\description{
Identifies songs that were most frequently played as opening songs in concerts,
excluding appearances as snippet as opening.
}
\examples{
\dontrun{
top_openers <- find_most_played_opening_songs(concert_data)
top_openers <- find_most_played_opening_songs(concert_data, n = 10)  # Show top 10
}

}
