% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{find_most_representative_setlists}
\alias{find_most_representative_setlists}
\title{Find Most Representative Setlists}
\usage{
find_most_representative_setlists(show_sequences_all, n_representatives = 30)
}
\arguments{
\item{show_sequences_all}{A data frame containing at least two columns:
\itemize{
  \item sequence: A character vector of setlist sequences.
  \item Other columns containing show information.
}}

\item{n_representatives}{An integer specifying the number of representative setlists to find. Default is 30.}
}
\value{
A data frame containing the most representative setlists,
        with the same structure as the input data frame.
}
\description{
This function identifies a specified number of shows that have the highest average
similarity to other shows' setlists, effectively finding the most representative
setlists in the dataset.
}
\examples{
\dontrun{
# Assuming show_sequences_all is your dataset
representative_shows <- find_most_representative_setlists(show_sequences_all, n_representatives = 30)
print(representative_shows)
}

}
