% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{find_representative_setlists_kmeans}
\alias{find_representative_setlists_kmeans}
\title{Find Most Representative Setlists using K-means Clustering}
\usage{
find_representative_setlists_kmeans(show_sequences_all, n_representatives = 30)
}
\arguments{
\item{show_sequences_all}{A data frame containing at least two columns:
\itemize{
  \item sequence: A character vector of setlist sequences.
  \item Other columns containing show information.
}}

\item{n_representatives}{An integer specifying the number of representative setlists to find. Default is 30.}
}
\value{
A data frame containing the most representative setlists,
        with the same structure as the input data frame.
}
\description{
This function identifies a specified number of shows that represent the variety
of setlists in the dataset, using k-means clustering.
}
\examples{
\dontrun{
# Assuming show_sequences_all is your dataset
representative_shows <- find_representative_setlists_kmeans(show_sequences_all, n_representatives = 30)
print(representative_shows)
}

}
