% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourStatistics.R
\name{find_shows_with_most_snippets}
\alias{find_shows_with_most_snippets}
\title{Find Shows with Most Snippets}
\usage{
find_shows_with_most_snippets(concert_data)
}
\arguments{
\item{concert_data}{A data frame containing concert information with columns:
date, city, country, venue, snippet (logical)}
}
\value{
A data frame sorted by number of snippets (descending) and date (descending),
        containing columns: date, city, country, venue, and number of snippets played
}
\description{
Identifies concerts with the highest number of song snippets played.
A snippet is a short piece of song (either from another artist or from U2 themselves)
played as part of a regular song in the show.
}
\examples{
\dontrun{
shows_with_snippets <- find_shows_with_most_snippets(concert_data)
head(shows_with_snippets)
}

}
