% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistAlignment.R
\name{four_letter_code_to_hex}
\alias{four_letter_code_to_hex}
\title{Convert 4-Letter Code to Unique Hex Character}
\usage{
four_letter_code_to_hex(code, used_hex_codes)
}
\arguments{
\item{code}{A character string of length 4 representing the song code.}

\item{used_hex_codes}{A vector of already used hex codes to ensure uniqueness.}
}
\value{
A list containing two elements:
  \itemize{
    \item char: A single character representing the unique hex character.
    \item value: The numeric value of the hex character.
  }
}
\description{
Convert 4-Letter Code to Unique Hex Character
}
