% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistVariability.R
\name{generate_setlist_data}
\alias{generate_setlist_data}
\title{Generate Setlist Data for Each Show}
\usage{
generate_setlist_data(data, tour_order = NULL)
}
\arguments{
\item{data}{A data frame in concertData format, containing at least 'showID', 'tour', 'date', and 'song_title' columns.}

\item{tour_order}{A data frame with a 'tour' column specifying the desired order of tours (optional).}
}
\value{
A data frame with columns:
  \itemize{
    \item showID: Unique identifier for each show
    \item tour: Name of the tour (as a factor if tour_order is provided)
    \item date: Date of the show
    \item songs: Number of songs played in the show
    \item song_list: Semicolon-separated list of all songs played in the show
  }
}
\description{
Creates a summary of setlist data for each show, including the number of unique songs and a list of all songs played.
}
\examples{
\dontrun{
setlist_summary <- generate_setlist_data(concert_data)
setlist_summary_ordered <- generate_setlist_data(concert_data, tour_order)
}

}
