% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourWorldMaps.R
\name{get_concert_locations}
\alias{get_concert_locations}
\title{Get City - Country Locations from Concert Data}
\usage{
get_concert_locations(data)
}
\arguments{
\item{data}{A data frame containing concert info with 'city' and 'country' columns.}
}
\value{
A data frame with unique city-country pairs and a combined 'location' column.
}
\description{
This function extracts city and country data from a data frame with concert info,
creating a unique list of locations that can for example be used to look up geographical coordinates
}
\examples{
city_data <- data.frame(
  city = c("New York", "London", ""),
  country = c("USA", "UK", "France")
)
unique_locations <- get_concert_locations(unique_shows)
}
