% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourSetlistVariability.R
\name{longest_setlist_by_tour}
\alias{longest_setlist_by_tour}
\title{Find Longest Setlist by Tour}
\usage{
longest_setlist_by_tour(data)
}
\arguments{
\item{data}{A data frame in concertData format, containing at least 'tour' and 'song_position' columns.}
}
\value{
A data frame with columns:
  \itemize{
    \item tour: Name of the tour
    \item max_songs: Maximum number of songs played in a single show for that tour
  }
}
\description{
Calculates the maximum number of songs played in a single show for each tour.
}
\examples{
\dontrun{
longest_setlists <- longest_setlist_by_tour(concert_data)
print(longest_setlists)
}

}
