% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tour_order_by_date}
\alias{tour_order_by_date}
\title{Order Tours by Date}
\usage{
tour_order_by_date(data, ascending = FALSE)
}
\arguments{
\item{data}{A data frame in concertData format, containing at least the following columns:
- `tour`: The name of the tour.
- `date`: The date of each concert in the tour.}

\item{ascending}{A logical value. If TRUE, tours are ordered from earliest to latest;
if FALSE (default), they are ordered from latest to earliest.}
}
\value{
A data frame with the following columns:
  - `tour`: A factor of tour names, ordered chronologically.
  - `median_date`: The median date of each tour.
}
\description{
This function orders tours based on their median date, allowing for either
             ascending or descending chronological order.
}
\examples{
\dontrun{
tour_order <- tour_order_by_date(concert_data)
tour_order_asc <- tour_order_by_date(concert_data, ascending = TRUE)
print(tour_order)
}

}
