% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tourStatistics.R
\name{find_longest_setlists}
\alias{find_longest_setlists}
\title{Find Longest Setlists}
\usage{
find_longest_setlists(concert_data)
}
\arguments{
\item{concert_data}{A data frame containing concert information with columns:
song_position, date, city, country, venue}
}
\value{
A data frame showing the concerts with the maximum number of songs,
        including date, city, country, venue, and number of songs played
}
\description{
This function identifies concerts with the maximum number of songs played
from a concert dataset.
}
\examples{
\dontrun{
data <- read.csv("u2concerts.csv")
find_longest_setlists(data)
}

}
